Added sell property

+++ Business_Sim/Building.cs
@@ -4,6 +4,10 @@
     {
         private int _upgradeLevel;
 
+        /// <summary>
+        /// Sets all default properties of the building depending on the type being added
+        /// </summary>
+        /// <param name="buildingType">Building type to add</param>
         public Building(BuildingType buildingType)
         {
             this.buildingType = buildingType;
@@ -79,8 +83,14 @@
             SkyScraper = 8
         }
 
+        /// <summary>
+        /// Building type of the building. Affects the base value for income, outcome, and buy/sell/upgrade price
+        /// </summary>
         public BuildingType buildingType { get; }
 
+        /// <summary>
+        /// Returns the building type as a string
+        /// </summary>
         public string buildingTypeString
         {
             get
@@ -124,11 +134,26 @@
             }
         }
 
+        /// <summary>
+        /// Price to buy/add the building to the owned buildings list
+        /// </summary>
         public decimal buyPrice { get; }
+        /// <summary>
+        /// Income building gives per game day
+        /// </summary>
         public decimal dailyIncome { get; private set; }
+        /// <summary>
+        /// Outcome building takes per game day
+        /// </summary>
         public decimal dailyOutcome { get; private set; }
+        /// <summary>
+        /// Amount of money gained if the building is sold/removed from the owned buildings list
+        /// </summary>
         public decimal sellPrice { get; private set; }
 
+        /// <summary>
+        /// Current upgrade level. Affects income, outcome, sell price and upgrade price.
+        /// </summary>
         public int upgradeLevel
         {
             get { return _upgradeLevel; }
@@ -137,11 +162,14 @@
                 _upgradeLevel = value;
                 dailyIncome = dailyIncome * _upgradeLevel;
                 dailyOutcome = dailyOutcome * _upgradeLevel;
-                sellPrice = (buyPrice / 2) * (_upgradeLevel / 2);
+                sellPrice = (buyPrice / 2) * ((decimal)_upgradeLevel / 2);
                 upgradePrice = buyPrice * (_upgradeLevel / 2);
             }
         }
 
+        /// <summary>
+        /// Price to increase the upgrade level of te building
+        /// </summary>
         public decimal upgradePrice { get; private set; }
     }
 }
